+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.2] - 2026-08-07
+
+### Fixed
+
+- Fixed downloads that failed while temporary files were still being written.
+- Fixed incomplete parallel download parts being reported too early.
+
 ## [2.1.1] - 2026-08-07
 
 ### Added